+++ network/src/trainer/training process/GD_momentum_process.h
@@ -20,7 +20,7 @@ namespace SNN_network
 	class GD_momentum_process : public Trainig_process
 	{
 	public:
-		GD_momentum_process(Perceptron* p_perceptron, double p_step, float momentum_factor);
+		GD_momentum_process(Perceptron* p_perceptron, double p_step, double momentum_factor);
 		~GD_momentum_process();
 
 		void propagate(vector<Trainig_process*> process, bool out = false);
@@ -29,7 +29,7 @@ namespace SNN_network
 
 	private:
 		double m_step;
-		float m_momentum_factor;
+		double m_momentum_factor;
 
 		double m_delta_1;
 	};
